feat(dataset): download dataset with async
Most of the performance are IO bound to the HTTP GET request emmited to the VitalDB API. We changed them into an async but we had to reimplement the `to_pandas` and `load_opendata` methods from VitalFile class of the vitaldb's. The current changes are not building a similar dataset. It should be resolved in future commits.

The performance is that we can retrieve the dataset under two minutes with a poor network VS former use of multiprocessing.

Notably, there are more works to perform:

- We do not get the meta data for each track type, reference to the vitaldb.utils:TRACK_INFO.
- We do not manage the sampling, start timestamp and end timestamp for each case. This point disables us to pass the `has_not_enough_data` flag on dataset_download:L240.
- We did not the result of the similarity of the downloaded dataset compared to the existint download_dataset script.
- The current timeout is not stable, on poor network, it might cancel download. We might fine tune the timeout or implement a loop to always try again on a TimeOutError.