chore(datasetbuilder): remove count of segments
Due to the use of multiprocessing, the state of the object should not be
stored. The state was updated with the number of correct segments and
number of segments considered. The state was not updated by a guard in
the multiprocessing lib, probably.

If we want to reimplement this info there we should use a mutex to avoid data race on these states